refactor(core): decompose Call into focused internal components
Break the ~2,260-line Call class into 11 internal collaborators under
call/components (CallApiClient, CallStatsReporter, CallRenderer,
CallEventManager, CallMediaManager, CallSessionManager,
CallIceConnectionMonitor, CallConnectivityMonitor, CallJoinCoordinator,
CallReconnector, CallLifecycleManager). Call remains a thin, binary-compatible
public facade that delegates to them; public API is unchanged (apiCheck passes).

Update white-box reflection tests to target the new component owners after
internals moved out of Call.
