+++ stream-video-android-core/src/main/kotlin/io/getstream/video/android/core/call/components/CallApiClient.kt
@@ -1,0 +1,402 @@
+/*
+ * Copyright (c) 2014-2026 Stream.io Inc. All rights reserved.
+ *
+ * Licensed under the Stream License;
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    https://github.com/GetStream/stream-video-android/blob/main/LICENSE
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.getstream.video.android.core.call.components
+
+import io.getstream.android.video.generated.models.AcceptCallResponse
+import io.getstream.android.video.generated.models.BlockUserResponse
+import io.getstream.android.video.generated.models.CallSettingsRequest
+import io.getstream.android.video.generated.models.GetCallResponse
+import io.getstream.android.video.generated.models.GetOrCreateCallResponse
+import io.getstream.android.video.generated.models.GoLiveResponse
+import io.getstream.android.video.generated.models.KickUserResponse
+import io.getstream.android.video.generated.models.ListRecordingsResponse
+import io.getstream.android.video.generated.models.ListTranscriptionsResponse
+import io.getstream.android.video.generated.models.MemberRequest
+import io.getstream.android.video.generated.models.MuteUsersResponse
+import io.getstream.android.video.generated.models.PinResponse
+import io.getstream.android.video.generated.models.RejectCallResponse
+import io.getstream.android.video.generated.models.RingCallRequest
+import io.getstream.android.video.generated.models.RingCallResponse
+import io.getstream.android.video.generated.models.SendCallEventResponse
+import io.getstream.android.video.generated.models.SendReactionResponse
+import io.getstream.android.video.generated.models.StartClosedCaptionsResponse
+import io.getstream.android.video.generated.models.StartTranscriptionResponse
+import io.getstream.android.video.generated.models.StopClosedCaptionsResponse
+import io.getstream.android.video.generated.models.StopLiveResponse
+import io.getstream.android.video.generated.models.StopTranscriptionResponse
+import io.getstream.android.video.generated.models.UnpinResponse
+import io.getstream.android.video.generated.models.UpdateCallMembersRequest
+import io.getstream.android.video.generated.models.UpdateCallMembersResponse
+import io.getstream.android.video.generated.models.UpdateCallRequest
+import io.getstream.android.video.generated.models.UpdateCallResponse
+import io.getstream.android.video.generated.models.UpdateUserPermissionsResponse
+import io.getstream.log.taggedLogger
+import io.getstream.result.Result
+import io.getstream.video.android.core.Call
+import io.getstream.video.android.core.RingingState
+import io.getstream.video.android.core.model.MuteUsersData
+import io.getstream.video.android.core.model.QueriedMembers
+import io.getstream.video.android.core.model.RejectReason
+import io.getstream.video.android.core.model.SortField
+import io.getstream.video.android.core.model.UpdateUserPermissionsData
+import io.getstream.video.android.core.recording.RecordingType
+import io.getstream.video.android.core.utils.toQueriedMembers
+import kotlinx.coroutines.launch
+import org.threeten.bp.OffsetDateTime
+
+/**
+ * Wraps all coordinator (REST) API calls for a [Call]. Each method delegates to the
+ * coordinator client and, where relevant, updates [Call.state] from the response.
+ *
+ * This component holds no mutable call state — it is a stateless façade over the
+ * coordinator endpoints, extracted from [Call] to keep the public class focused.
+ */
+internal class CallApiClient(
+    private val call: Call,
+) {
+    private val logger by taggedLogger("Call:ApiClient:${call.type}:${call.id}")
+
+    private val clientImpl get() = call.clientImpl
+    private val type get() = call.type
+    private val id get() = call.id
+    private val state get() = call.state
+
+    suspend fun get(): Result<GetCallResponse> {
+        val response = clientImpl.getCall(type, id)
+        response.onSuccess {
+            state.updateFromResponse(it)
+        }
+        return response
+    }
+
+    suspend fun create(
+        memberIds: List<String>? = null,
+        members: List<MemberRequest>? = null,
+        custom: Map<String, Any>? = null,
+        settings: CallSettingsRequest? = null,
+        startsAt: OffsetDateTime? = null,
+        team: String? = null,
+        ring: Boolean = false,
+        notify: Boolean = false,
+        video: Boolean? = null,
+    ): Result<GetOrCreateCallResponse> {
+        val response = if (members != null) {
+            clientImpl.getOrCreateCallFullMembers(
+                type = type,
+                id = id,
+                members = members,
+                custom = custom,
+                settingsOverride = settings,
+                startsAt = startsAt,
+                team = team,
+                ring = ring,
+                notify = notify,
+                video = video,
+            )
+        } else {
+            clientImpl.getOrCreateCall(
+                type = type,
+                id = id,
+                memberIds = memberIds,
+                custom = custom,
+                settingsOverride = settings,
+                startsAt = startsAt,
+                team = team,
+                ring = ring,
+                notify = notify,
+                video = video,
+            )
+        }
+
+        response.onSuccess {
+            /**
+             * Because [io.getstream.video.android.core.CallState.updateFromResponse] reads the
+             * value of [io.getstream.video.android.core.ClientState.ringingCall]
+             */
+            if (ring) {
+                call.client.state._ringingCall.value = call
+            }
+            state.updateFromResponse(it)
+            if (ring) {
+                call.client.state.addRingingCall(call, RingingState.Outgoing())
+            }
+        }
+        return response
+    }
+
+    suspend fun update(
+        custom: Map<String, Any>? = null,
+        settingsOverride: CallSettingsRequest? = null,
+        startsAt: OffsetDateTime? = null,
+    ): Result<UpdateCallResponse> {
+        val request = UpdateCallRequest(
+            custom = custom,
+            settingsOverride = settingsOverride,
+            startsAt = startsAt,
+        )
+        val response = clientImpl.updateCall(type, id, request)
+        response.onSuccess {
+            state.updateFromResponse(it)
+        }
+        return response
+    }
+
+    suspend fun pinForEveryone(sessionId: String, userId: String): Result<PinResponse> {
+        return clientImpl.pinForEveryone(type, id, sessionId, userId)
+    }
+
+    suspend fun unpinForEveryone(sessionId: String, userId: String): Result<UnpinResponse> {
+        return clientImpl.unpinForEveryone(type, id, sessionId, userId)
+    }
+
+    suspend fun sendReaction(
+        type: String,
+        emoji: String? = null,
+        custom: Map<String, Any>? = null,
+    ): Result<SendReactionResponse> {
+        return clientImpl.sendReaction(this.type, id, type, emoji, custom)
+    }
+
+    suspend fun queryMembers(
+        filter: Map<String, Any>,
+        sort: List<SortField> = mutableListOf(SortField.Desc("created_at")),
+        limit: Int = 25,
+        prev: String? = null,
+        next: String? = null,
+    ): Result<QueriedMembers> {
+        return clientImpl.queryMembersInternal(
+            type = type,
+            id = id,
+            filter = filter,
+            sort = sort,
+            prev = prev,
+            next = next,
+            limit = limit,
+        ).onSuccess { state.updateFromResponse(it) }.map { it.toQueriedMembers() }
+    }
+
+    suspend fun muteAllUsers(
+        audio: Boolean = true,
+        video: Boolean = false,
+        screenShare: Boolean = false,
+    ): Result<MuteUsersResponse> {
+        val request = MuteUsersData(
+            muteAllUsers = true,
+            audio = audio,
+            video = video,
+            screenShare = screenShare,
+        )
+        return clientImpl.muteUsers(type, id, request)
+    }
+
+    suspend fun goLive(
+        startHls: Boolean = false,
+        startRecording: Boolean = false,
+        startTranscription: Boolean = false,
+    ): Result<GoLiveResponse> {
+        val result = clientImpl.goLive(
+            type = type,
+            id = id,
+            startHls = startHls,
+            startRecording = startRecording,
+            startTranscription = startTranscription,
+        )
+        result.onSuccess { state.updateFromResponse(it) }
+
+        return result
+    }
+
+    suspend fun stopLive(): Result<StopLiveResponse> {
+        val result = clientImpl.stopLive(type, id)
+        result.onSuccess { state.updateFromResponse(it) }
+        return result
+    }
+
+    suspend fun sendCustomEvent(data: Map<String, Any>): Result<SendCallEventResponse> {
+        return clientImpl.sendCustomEvent(type, id, data)
+    }
+
+    suspend fun requestPermissions(vararg permission: String): Result<Unit> {
+        return clientImpl.requestPermissions(type, id, permission.toList())
+    }
+
+    suspend fun startRecording(recordingType: RecordingType): Result<Any> {
+        return clientImpl.startRecording(type, id, recordingType = recordingType)
+    }
+
+    suspend fun stopRecording(recordingType: RecordingType): Result<Any> {
+        return clientImpl.stopRecording(type, id, recordingType)
+    }
+
+    suspend fun startHLS(): Result<Any> {
+        return clientImpl.startBroadcasting(type, id)
+            .onSuccess {
+                state.updateFromResponse(it)
+            }
+    }
+
+    suspend fun stopHLS(): Result<Any> {
+        return clientImpl.stopBroadcasting(type, id)
+    }
+
+    suspend fun blockUser(userId: String): Result<BlockUserResponse> {
+        return clientImpl.blockUser(type, id, userId)
+    }
+
+    suspend fun removeMembers(userIds: List<String>): Result<UpdateCallMembersResponse> {
+        val request = UpdateCallMembersRequest(removeMembers = userIds)
+        return clientImpl.updateMembers(type, id, request)
+    }
+
+    suspend fun grantPermissions(
+        userId: String,
+        permissions: List<String>,
+    ): Result<UpdateUserPermissionsResponse> {
+        val request = UpdateUserPermissionsData(
+            userId = userId,
+            grantedPermissions = permissions,
+        )
+        return clientImpl.updateUserPermissions(type, id, request)
+    }
+
+    suspend fun revokePermissions(
+        userId: String,
+        permissions: List<String>,
+    ): Result<UpdateUserPermissionsResponse> {
+        val request = UpdateUserPermissionsData(
+            userId = userId,
+            revokedPermissions = permissions,
+        )
+        return clientImpl.updateUserPermissions(type, id, request)
+    }
+
+    suspend fun updateMembers(memberRequests: List<MemberRequest>): Result<UpdateCallMembersResponse> {
+        val request = UpdateCallMembersRequest(updateMembers = memberRequests)
+        return clientImpl.updateMembers(type, id, request)
+    }
+
+    suspend fun listRecordings(sessionId: String? = null): Result<ListRecordingsResponse> {
+        return clientImpl.listRecordings(type, id, sessionId)
+    }
+
+    suspend fun kickUser(
+        userId: String,
+        block: Boolean = false,
+    ): Result<KickUserResponse> = clientImpl.kickUser(
+        type,
+        id,
+        userId,
+        block,
+    )
+
+    suspend fun muteUser(
+        userId: String,
+        audio: Boolean = true,
+        video: Boolean = false,
+        screenShare: Boolean = false,
+    ): Result<MuteUsersResponse> {
+        val request = MuteUsersData(
+            users = listOf(userId),
+            muteAllUsers = false,
+            audio = audio,
+            video = video,
+            screenShare = screenShare,
+        )
+        return clientImpl.muteUsers(type, id, request)
+    }
+
+    suspend fun muteUsers(
+        userIds: List<String>,
+        audio: Boolean = true,
+        video: Boolean = false,
+        screenShare: Boolean = false,
+    ): Result<MuteUsersResponse> {
+        val request = MuteUsersData(
+            users = userIds,
+            muteAllUsers = false,
+            audio = audio,
+            video = video,
+            screenShare = screenShare,
+        )
+        return clientImpl.muteUsers(type, id, request)
+    }
+
+    suspend fun ring(): Result<GetCallResponse> {
+        logger.d { "[ring] #ringing; no args" }
+        return clientImpl.ring(type, id)
+    }
+
+    suspend fun ring(ringCallRequest: RingCallRequest): Result<RingCallResponse> {
+        logger.d { "[ring] #ringing ringCallRequest: $ringCallRequest" }
+        return clientImpl.ring(type, id, ringCallRequest)
+    }
+
+    suspend fun notify(): Result<GetCallResponse> {
+        logger.d { "[notify] #ringing; no args" }
+        return clientImpl.notify(type, id)
+    }
+
+    suspend fun accept(): Result<AcceptCallResponse> {
+        logger.d { "[accept] #ringing; no args, call_id:$id" }
+        state.acceptedOnThisDevice = true
+
+        clientImpl.state.transitionToAcceptCall(call)
+        return clientImpl.accept(type, id)
+    }
+
+    suspend fun reject(reason: RejectReason? = null): Result<RejectCallResponse> {
+        logger.d { "[reject] #ringing; rejectReason: $reason, call_id:$id" }
+        return clientImpl.reject(type, id, reason)
+    }
+
+    fun collectUserFeedback(
+        rating: Int,
+        reason: String? = null,
+        custom: Map<String, Any>? = null,
+    ) {
+        call.scope.launch {
+            clientImpl.collectFeedback(
+                callType = type,
+                id = id,
+                sessionId = call.sessionId,
+                rating = rating,
+                reason = reason,
+                custom = custom,
+            )
+        }
+    }
+
+    suspend fun startTranscription(): Result<StartTranscriptionResponse> {
+        return clientImpl.startTranscription(type, id)
+    }
+
+    suspend fun stopTranscription(): Result<StopTranscriptionResponse> {
+        return clientImpl.stopTranscription(type, id)
+    }
+
+    suspend fun listTranscription(): Result<ListTranscriptionsResponse> {
+        return clientImpl.listTranscription(type, id)
+    }
+
+    suspend fun startClosedCaptions(): Result<StartClosedCaptionsResponse> {
+        return clientImpl.startClosedCaptions(type, id)
+    }
+
+    suspend fun stopClosedCaptions(): Result<StopClosedCaptionsResponse> {
+        return clientImpl.stopClosedCaptions(type, id)
+    }
+}
